Updating hash tree code

+++ src/java/voldemort/hashtrees/SegmentHash.java
@@ -14,7 +14,7 @@ public class SegmentHash {
         return hash;
     }
 
-    public long getNodeId() {
+    public int getNodeId() {
         return nodeId;
     }
